Rename validation module for bridge client
Related to: #268
