+++ tools/bridge/bridge/contract_validation.py
